refactor: commenting code that displays images in window
encountered errors similar to the one below:

cv2.error: OpenCV(3.4.2) 
/tmp/build/80754af9/opencv-suite_1535558631802/work/modules/highgui/src/window.cpp:698: 
error: (-2:Unspecified error) The function is not i
mplemented. Rebuild the library with Windows, GTK+ 2.x or Carbon 
support. If you are on Ubuntu or Debian, install libgtk2.0-dev and 
pkg-config, then re-run cmake or
 configure script in function 'cvWaitKey'

To try upgrading opencv as suggested here - 
https://stackoverflow.com/a/14656610/315385 - when ubuntu is upgraded to 
18.04